ver1.1

+++ Post/views.py
@@ -1,13 +1,32 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Post, Like
-from .forms import PostForm
+from .models import Post, Like, Comment
+from .forms import PostForm, CommentForm
 from django.db import IntegrityError
 from django.contrib import messages
 
 
+def comment_adjustment(request, id, id2):
+    obj = get_object_or_404(Comment, id=id)
+    post = get_object_or_404(Post, id=id2)
+    form = CommentForm(request.POST or None, instance=obj)
+    context = {'obj': obj, 'post': post}
+
+    if form.is_valid():
+        form.save()
+        return redirect('/')
+
+    context['form'] = form
+
+    if request.method == 'POST' and 'c_delete' in request.POST:
+        obj.delete()
+        return redirect('/')
+
+    return render(request, 'posts/comment_delete.html', context)
+
+
 def like_view(request, id):
     post = Post.objects.get(id=id)
-    likes = post.post.all() # Getting all users that liked this post
+    likes = post.post.all()  # Getting all users that liked this post
 
     context = {
         'likes': likes
@@ -16,6 +35,7 @@ def like_view(request, id):
 
 
 def detail_view(request, id):
+    user = request.user
     post = Post.objects.get(id=id)
     like_count = post.post.filter().values('like').count()
     likes = post.post.all().filter().values('like')
@@ -26,21 +46,34 @@ def detail_view(request, id):
 
     try:
         if like:
-            if request.method == 'POST':
+            if request.method == 'POST' and 'likeform' in request.POST:
                 Like.objects.create(like=request.user, post=post)
                 return redirect('/p/%d' % post.id)
         else:
-            if request.method == 'POST':
+            if request.method == 'POST' and 'likeform' in request.POST:
                 Like.objects.filter(like=request.user, post=post).delete()
                 return redirect('/p/%d' % post.id)
 
     except IntegrityError:
         messages.warning(request, 'no')
 
+    form = CommentForm()
+    comments = post.c_post.all()
+
+    if request.method == 'POST' and 'commentform' in request.POST:
+        form = CommentForm(request.POST or None)
+        form.instance.c_post = post
+        form.instance.user = user
+        if form.is_valid():
+            form.save()
+            return redirect('/p/%d' % post.id)
+
     context = {
         'post': post,
         'like_count': like_count,
-        'like': like
+        'like': like,
+        'form': form,
+        'comments': comments,
     }
 
     return render(request, 'posts/post_index.html', context)
